Refactor app startup
