+++ lib/app/app.dart
@@ -1,0 +1,3 @@
+export 'app_theme_manager/app_theme_manager.dart';
+export 'routes/routes.dart';
+export 'view/app.dart';
