+++ lib/main/main.dart
@@ -1,0 +1,6 @@
+import 'package:l_breez/app/view/app.dart';
+import 'package:l_breez/main/bootstrap.dart';
+
+void main() {
+  bootstrap((injector) => App(injector: injector));
+}
